+++ docs/changelog.md
@@ -17,6 +17,7 @@ MML changes:
  * Added `P` pitches to portamento
  * Fixed a state tracking bug involving loops that can cause dropped instructions.
  * Fixed missing `set_detune 0` instruction when `MD` value is small
+ * Fixed MP and portamento not accounting for manual `D` detune
 
 GUI changes:
  * Sample Analyser can now analyse samples up to a length of 32768.